+++ packages/backend-rs/src/model/entity/sea_orm_active_enums.rs
@@ -128,6 +128,22 @@ pub enum RelayStatusEnum {
 #[sea_orm(
     rs_type = "String",
     db_type = "Enum",
+    enum_name = "user_emojimodperm_enum"
+)]
+pub enum UserEmojimodpermEnum {
+    #[sea_orm(string_value = "add")]
+    Add,
+    #[sea_orm(string_value = "full")]
+    Full,
+    #[sea_orm(string_value = "mod")]
+    Mod,
+    #[sea_orm(string_value = "unauthorized")]
+    Unauthorized,
+}
+#[derive(Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum)]
+#[sea_orm(
+    rs_type = "String",
+    db_type = "Enum",
     enum_name = "user_profile_ffvisibility_enum"
 )]
 pub enum UserProfileFfvisibilityEnum {