+++ packages/backend/src/server/api/endpoints/admin/show-user.ts
@@ -65,6 +65,7 @@ export default define(meta, paramDef, async (ps, me) => {
 		isModerator: user.isModerator,
 		isSilenced: user.isSilenced,
 		isSuspended: user.isSuspended,
+		emojiModPerm: user.emojiModPerm,
 		lastActiveDate: user.lastActiveDate,
 		moderationNote: profile.moderationNote,
 		signins,